requirements.txt
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
Pillow==10.2.0
PyYAML==6.0.1
torch==1.8.0+cu111
torchvision==0.9.0+cu111
